Security Update & Bugfix
Call data type was incorrectly set to bytes32 instead of bytes.

Added check to ensure that the origin of a call to
VouchersUser.forwardRedeemedVouchers is the registry, since calling an
arbitrary function on an arbitrary contract is a huge security risk.

Began updating code to be compatible with solidity 0.5.x